+++ lxd/container_snapshot.go
@@ -1,0 +1,276 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"os"
+	"os/exec"
+	"path"
+	"strconv"
+
+	"github.com/gorilla/mux"
+	"gopkg.in/lxc/go-lxc.v2"
+
+	"github.com/lxc/lxd/shared"
+)
+
+func snapshotsDir(c *lxdContainer) string {
+	return shared.VarPath("lxc", c.name, "snapshots")
+}
+
+func snapshotDir(c *lxdContainer, name string) string {
+	return path.Join(snapshotsDir(c), name)
+}
+
+func snapshotStateDir(c *lxdContainer, name string) string {
+	return path.Join(snapshotDir(c, name), "state")
+}
+
+func snapshotRootfsDir(c *lxdContainer, name string) string {
+	return path.Join(snapshotDir(c, name), "rootfs")
+}
+
+func containerSnapshotsGet(d *Daemon, r *http.Request) Response {
+	recursion_str := r.FormValue("recursion")
+	recursion, err := strconv.Atoi(recursion_str)
+	if err != nil {
+		recursion = 0
+	}
+
+	cname := mux.Vars(r)["name"]
+	c, err := newLxdContainer(cname, d)
+	if err != nil {
+		return SmartError(err)
+	}
+
+	regexp := fmt.Sprintf("%s/", cname)
+	length := len(regexp)
+	q := "SELECT name FROM containers WHERE type=? AND SUBSTR(name,1,?)=?"
+	var name string
+	inargs := []interface{}{cTypeSnapshot, length, regexp}
+	outfmt := []interface{}{name}
+	results, err := shared.DbQueryScan(d.db, q, inargs, outfmt)
+	if err != nil {
+		return SmartError(err)
+	}
+
+	var result_string []string
+	var result_map []shared.Jmap
+
+	for _, r := range results {
+		name = r[0].(string)
+		if recursion == 0 {
+			url := fmt.Sprintf("/%s/containers/%s/snapshots/%s", shared.APIVersion, cname, name)
+			result_string = append(result_string, url)
+		} else {
+			_, err := os.Stat(snapshotStateDir(c, name))
+			body := shared.Jmap{"name": name, "stateful": err == nil}
+			result_map = append(result_map, body)
+		}
+
+	}
+
+	if recursion == 0 {
+		return SyncResponse(true, result_string)
+	} else {
+		return SyncResponse(true, result_map)
+	}
+
+}
+
+/*
+ * Note, the code below doesn't deal with snapshots of snapshots.
+ * To do that, we'll need to weed out based on # slashes in names
+ */
+func nextSnapshot(d *Daemon, name string) int {
+	base := fmt.Sprintf("%s/snap", name)
+	length := len(base)
+	q := fmt.Sprintf("SELECT MAX(name) FROM containers WHERE type=? AND SUBSTR(name,1,?)=?")
+	var numstr string
+	inargs := []interface{}{cTypeSnapshot, length, base}
+	outfmt := []interface{}{numstr}
+	results, err := shared.DbQueryScan(d.db, q, inargs, outfmt)
+	if err != nil {
+		return 0
+	}
+	max := 0
+
+	for _, r := range results {
+		numstr = r[0].(string)
+		if len(numstr) <= length {
+			continue
+		}
+		substr := numstr[length:]
+		var num int
+		count, err := fmt.Sscanf(substr, "%d", &num)
+		if err != nil || count != 1 {
+			continue
+		}
+		if num >= max {
+			max = num + 1
+		}
+	}
+
+	return max
+}
+
+func containerSnapshotsPost(d *Daemon, r *http.Request) Response {
+	name := mux.Vars(r)["name"]
+
+	/*
+	 * snapshot is a three step operation:
+	 * 1. choose a new name
+	 * 2. copy the database info over
+	 * 3. copy over the rootfs
+	 */
+	c, err := newLxdContainer(name, d)
+	if err != nil {
+		return SmartError(err)
+	}
+
+	raw := shared.Jmap{}
+	if err := json.NewDecoder(r.Body).Decode(&raw); err != nil {
+		return BadRequest(err)
+	}
+
+	snapshotName, err := raw.GetString("name")
+	if err != nil || snapshotName == "" {
+		// come up with a name
+		i := nextSnapshot(d, name)
+		snapshotName = fmt.Sprintf("snap%d", i)
+	}
+
+	stateful, err := raw.GetBool("stateful")
+	if err != nil {
+		return BadRequest(err)
+	}
+
+	fullName := fmt.Sprintf("%s/%s", name, snapshotName)
+	snapDir := snapshotDir(c, snapshotName)
+	if shared.PathExists(snapDir) {
+		return Conflict
+	}
+
+	err = os.MkdirAll(snapDir, 0700)
+	if err != nil {
+		return InternalError(err)
+	}
+
+	snapshot := func() error {
+
+		StateDir := snapshotStateDir(c, snapshotName)
+		err = os.MkdirAll(StateDir, 0700)
+		if err != nil {
+			return err
+		}
+
+		if stateful {
+			// TODO - shouldn't we freeze for the duration of rootfs snapshot below?
+			if !c.c.Running() {
+				return fmt.Errorf("Container not running\n")
+			}
+			opts := lxc.CheckpointOptions{Directory: StateDir, Stop: true, Verbose: true}
+			if err := c.c.Checkpoint(opts); err != nil {
+				return err
+			}
+		}
+
+		/* Create the db info */
+		args := DbCreateContainerArgs{
+			d:            d,
+			name:         fullName,
+			ctype:        cTypeSnapshot,
+			config:       c.config,
+			profiles:     c.profiles,
+			ephem:        c.ephemeral,
+			baseImage:    c.config["volatile.baseImage"],
+			architecture: c.architecture,
+		}
+
+		_, err := dbCreateContainer(args)
+		if err != nil {
+			return err
+		}
+
+		/* Create the directory and rootfs, set perms */
+		/* Copy the rootfs */
+		oldPath := fmt.Sprintf("%s/", shared.VarPath("lxc", name, "rootfs"))
+		newPath := snapshotRootfsDir(c, snapshotName)
+		err = exec.Command("rsync", "-a", "--devices", oldPath, newPath).Run()
+		return err
+	}
+
+	return AsyncResponse(shared.OperationWrap(snapshot), nil)
+}
+
+func snapshotHandler(d *Daemon, r *http.Request) Response {
+	containerName := mux.Vars(r)["name"]
+	c, err := newLxdContainer(containerName, d)
+	if err != nil {
+		return SmartError(err)
+	}
+
+	snapshotName := mux.Vars(r)["snapshotName"]
+	dir := snapshotDir(c, snapshotName)
+
+	_, err = os.Stat(dir)
+	if err != nil {
+		return SmartError(err)
+	}
+
+	switch r.Method {
+	case "GET":
+		return snapshotGet(c, snapshotName)
+	case "POST":
+		return snapshotPost(r, c, snapshotName)
+	case "DELETE":
+		return snapshotDelete(d, c, snapshotName)
+	default:
+		return NotFound
+	}
+}
+
+func snapshotGet(c *lxdContainer, name string) Response {
+	_, err := os.Stat(snapshotStateDir(c, name))
+	body := shared.Jmap{"name": name, "stateful": err == nil}
+	return SyncResponse(true, body)
+}
+
+func snapshotPost(r *http.Request, c *lxdContainer, oldName string) Response {
+	raw := shared.Jmap{}
+	if err := json.NewDecoder(r.Body).Decode(&raw); err != nil {
+		return BadRequest(err)
+	}
+
+	newName, err := raw.GetString("name")
+	if err != nil {
+		return BadRequest(err)
+	}
+
+	oldDir := snapshotDir(c, oldName)
+	newDir := snapshotDir(c, newName)
+
+	_, err = os.Stat(newDir)
+	if !os.IsNotExist(err) {
+		return InternalError(err)
+	} else if err == nil {
+		return Conflict
+	}
+
+	/*
+	 * TODO: do we need to do something more intelligent here? We probably
+	 * shouldn't do anything for stateful snapshots, since changing the fs
+	 * out from under criu will cause it to fail, but it may be useful to
+	 * do something for stateless ones.
+	 */
+	rename := func() error { return os.Rename(oldDir, newDir) }
+	return AsyncResponse(shared.OperationWrap(rename), nil)
+}
+
+func snapshotDelete(d *Daemon, c *lxdContainer, name string) Response {
+	dbRemoveSnapshot(d, c.name, name)
+	dir := snapshotDir(c, name)
+	remove := func() error { return os.RemoveAll(dir) }
+	return AsyncResponse(shared.OperationWrap(remove), nil)
+}
