+++ lxd/containers.go
@@ -1,134 +1,77 @@
 package main
 
 import (
-	"bytes"
-	"crypto/rand"
-	"database/sql"
-	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"math/big"
-	"net"
-	"net/http"
 	"os"
 	"os/exec"
-	"path"
-	"path/filepath"
-	"regexp"
 	"runtime"
-	"sort"
-	"strconv"
-	"strings"
 	"sync"
-	"syscall"
 	"time"
 
-	"github.com/dustinkirkland/golang-petname"
-	"github.com/gorilla/mux"
 	"github.com/gorilla/websocket"
 	"gopkg.in/lxc/go-lxc.v2"
 
-	"github.com/lxc/lxd/lxd/migration"
 	"github.com/lxc/lxd/shared"
 )
 
-type containerType int
-
-const (
-	cTypeRegular  containerType = 0
-	cTypeSnapshot containerType = 1
-)
-
-func containersGet(d *Daemon, r *http.Request) Response {
-	for {
-		result, err := doContainersGet(d, d.isRecursionRequest(r))
-		if err == nil {
-			return SyncResponse(true, result)
-		}
-		if !shared.IsDbLockedError(err) {
-			shared.Debugf("DBERR: containersGet: error %q\n", err)
-			return InternalError(err)
-		}
-		// 1 s may seem drastic, but we really don't want to thrash
-		// perhaps we should use a random amount
-		shared.Debugf("DBERR: containersGet, db is locked\n")
-		shared.PrintStack()
-		time.Sleep(1 * time.Second)
-	}
+type lxdContainer struct {
+	c            *lxc.Container
+	daemon       *Daemon
+	id           int
+	name         string
+	config       map[string]string
+	profiles     []string
+	devices      shared.Devices
+	architecture int
+	ephemeral    bool
+	idmapset     *shared.IdmapSet
 }
 
-func doContainerGet(d *Daemon, cname string) (shared.ContainerInfo, Response) {
-	_, err := dbGetContainerId(d.db, cname)
-	if err != nil {
-		return shared.ContainerInfo{}, SmartError(err)
-	}
-
-	c, err := newLxdContainer(cname, d)
-	if err != nil {
-		return shared.ContainerInfo{}, SmartError(err)
-	}
-
-	var name string
-	regexp := fmt.Sprintf("%s/", cname)
-	length := len(regexp)
-	q := "SELECT name FROM containers WHERE type=? AND SUBSTR(name,1,?)=?"
-	inargs := []interface{}{cTypeSnapshot, length, regexp}
-	outfmt := []interface{}{name}
-	results, err := shared.DbQueryScan(d.db, q, inargs, outfmt)
-	if err != nil {
-		return shared.ContainerInfo{}, SmartError(err)
-	}
-
-	var body []string
-
-	for _, r := range results {
-		name = r[0].(string)
-
-		url := fmt.Sprintf("/%s/containers/%s/snapshots/%s", shared.APIVersion, cname, name)
-		body = append(body, url)
-	}
-
-	cts, err := c.RenderState()
-	if err != nil {
-		return shared.ContainerInfo{}, SmartError(err)
-	}
-
-	containerinfo := shared.ContainerInfo{State: *cts,
-		Snaps: body}
-
-	return containerinfo, nil
+type execWs struct {
+	command          []string
+	container        *lxc.Container
+	rootUid          int
+	rootGid          int
+	options          lxc.AttachOptions
+	conns            map[int]*websocket.Conn
+	allConnected     chan bool
+	controlConnected chan bool
+	interactive      bool
+	done             chan shared.OperationResult
+	fds              map[int]string
 }
 
-func doContainersGet(d *Daemon, recursion bool) (interface{}, error) {
-	result, err := dbListContainers(d)
-	if err != nil {
-		return nil, err
-	}
+type commandPostContent struct {
+	Command     []string          `json:"command"`
+	WaitForWS   bool              `json:"wait-for-websocket"`
+	Interactive bool              `json:"interactive"`
+	Environment map[string]string `json:"environment"`
+}
 
-	result_string := make([]string, 0)
-	result_map := make([]shared.ContainerInfo, 0)
-	if err != nil {
-		return []string{}, err
-	}
-	for _, container := range result {
-		if !recursion {
-			url := fmt.Sprintf("/%s/containers/%s", shared.APIVersion, container)
-			result_string = append(result_string, url)
-		} else {
-			container, response := doContainerGet(d, container)
-			if response != nil {
-				continue
-			}
-			result_map = append(result_map, container)
-		}
-	}
+type containerConfigReq struct {
+	Profiles []string          `json:"profiles"`
+	Config   map[string]string `json:"config"`
+	Devices  shared.Devices    `json:"devices"`
+	Restore  string            `json:"restore"`
+}
 
-	if !recursion {
-		return result_string, nil
-	} else {
-		return result_map, nil
-	}
+type containerStatePutReq struct {
+	Action  string `json:"action"`
+	Timeout int    `json:"timeout"`
+	Force   bool   `json:"force"`
+}
+
+type containerPostBody struct {
+	Migration bool   `json:"migration"`
+	Name      string `json:"name"`
+}
+
+type containerPostReq struct {
+	Name      string               `json:"name"`
+	Source    containerImageSource `json:"source"`
+	Config    map[string]string    `json:"config"`
+	Profiles  []string             `json:"profiles"`
+	Ephemeral bool                 `json:"ephemeral"`
 }
 
 type containerImageSource struct {
@@ -158,12 +101,48 @@ type containerImageSource struct {
 	Source string `json:"source"`
 }
 
-type containerPostReq struct {
-	Name      string               `json:"name"`
-	Source    containerImageSource `json:"source"`
-	Config    map[string]string    `json:"config"`
-	Profiles  []string             `json:"profiles"`
-	Ephemeral bool                 `json:"ephemeral"`
+var containersCmd = Command{
+	name: "containers",
+	get:  containersGet,
+	post: containersPost,
+}
+
+var containerCmd = Command{
+	name:   "containers/{name}",
+	get:    containerGet,
+	put:    containerPut,
+	delete: containerDelete,
+	post:   containerPost,
+}
+
+var containerStateCmd = Command{
+	name: "containers/{name}/state",
+	get:  containerStateGet,
+	put:  containerStatePut,
+}
+
+var containerFileCmd = Command{
+	name: "containers/{name}/files",
+	get:  containerFileHandler,
+	post: containerFileHandler,
+}
+
+var containerSnapshotsCmd = Command{
+	name: "containers/{name}/snapshots",
+	get:  containerSnapshotsGet,
+	post: containerSnapshotsPost,
+}
+
+var containerSnapshotCmd = Command{
+	name:   "containers/{name}/snapshots/{snapshotName}",
+	get:    snapshotHandler,
+	post:   snapshotHandler,
+	delete: snapshotHandler,
+}
+
+var containerExecCmd = Command{
+	name: "containers/{name}/exec",
+	post: containerExecPost,
 }
 
 func containerWatchEphemeral(c *lxdContainer) {
@@ -283,717 +262,6 @@ func containersShutdown(d *Daemon) error {
 	return nil
 }
 
-func createFromImage(d *Daemon, req *containerPostReq) Response {
-	var hash string
-	var err error
-	var run func() shared.OperationResult
-
-	backing_fs, err := shared.GetFilesystem(d.lxcpath)
-	if err != nil {
-		return InternalError(err)
-	}
-
-	if req.Source.Alias != "" {
-		if req.Source.Mode == "pull" && req.Source.Server != "" {
-			hash, err = remoteGetImageFingerprint(d, req.Source.Server, req.Source.Alias)
-			if err != nil {
-				return InternalError(err)
-			}
-		} else {
-
-			hash, err = dbAliasGet(d.db, req.Source.Alias)
-			if err != nil {
-				return InternalError(err)
-			}
-		}
-	} else if req.Source.Fingerprint != "" {
-		hash = req.Source.Fingerprint
-	} else {
-		return BadRequest(fmt.Errorf("must specify one of alias or fingerprint for init from image"))
-	}
-
-	if req.Source.Server != "" {
-		err := ensureLocalImage(d, req.Source.Server, hash, req.Source.Secret)
-		if err != nil {
-			return InternalError(err)
-		}
-	}
-
-	imgInfo, err := dbImageGet(d.db, hash, false)
-	if err != nil {
-		return SmartError(err)
-	}
-	hash = imgInfo.Fingerprint
-
-	dpath := shared.VarPath("lxc", req.Name)
-	if shared.PathExists(dpath) {
-		return InternalError(fmt.Errorf("Container exists"))
-	}
-
-	name := req.Name
-
-	args := DbCreateContainerArgs{
-		d:            d,
-		name:         name,
-		ctype:        cTypeRegular,
-		config:       req.Config,
-		profiles:     req.Profiles,
-		ephem:        req.Ephemeral,
-		baseImage:    hash,
-		architecture: imgInfo.Architecture,
-	}
-
-	_, err = dbCreateContainer(args)
-	if err != nil {
-		removeContainerPath(d, name)
-		return SmartError(err)
-	}
-
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		removeContainer(d, name)
-		return SmartError(err)
-	}
-
-	vgname, vgnameIsSet, err := getServerConfigValue(d, "core.lvm_vg_name")
-	if err != nil {
-		return InternalError(fmt.Errorf("Error checking server config: %v", err))
-	}
-
-	if vgnameIsSet && shared.PathExists(fmt.Sprintf("%s.lv", shared.VarPath("images", hash))) {
-		run = shared.OperationWrap(func() error {
-			srcLVPath := fmt.Sprintf("/dev/%s/%s", vgname, hash)
-
-			snapshotLVPath, err := shared.LVMCreateSnapshotLV(name, hash, vgname)
-			if err != nil {
-				return fmt.Errorf("Error creating snapshot of source volume '%s'", srcLVPath)
-			}
-
-			destPath := shared.VarPath("lxc", name)
-			err = os.MkdirAll(destPath, 0700)
-			if err != nil {
-				return fmt.Errorf("Error creating container directory: %v", err)
-			}
-
-			output, err := exec.Command("mount", "-o", "discard", snapshotLVPath, destPath).CombinedOutput()
-			if err != nil {
-				return fmt.Errorf("Error mounting snapshot LV: %v\noutput:'%s'", err, output)
-			}
-
-			if !c.isPrivileged() {
-				return shiftRootfs(c, name, d)
-			}
-
-			return nil
-		})
-
-	} else if backing_fs == "btrfs" && shared.PathExists(fmt.Sprintf("%s.btrfs", shared.VarPath("images", hash))) {
-		run = shared.OperationWrap(func() error {
-			if err := btrfsCopyImage(hash, name, d); err != nil {
-				return err
-			}
-
-			if !c.isPrivileged() {
-				err = shiftRootfs(c, name, d)
-				if err != nil {
-					return err
-				}
-			}
-
-			err = templateApply(c, "create")
-			if err != nil {
-				return err
-			}
-
-			return nil
-		})
-
-	} else {
-		rootfsPath := fmt.Sprintf("%s/rootfs", dpath)
-		err = os.MkdirAll(rootfsPath, 0700)
-		if err != nil {
-			return InternalError(fmt.Errorf("Error creating rootfs directory"))
-		}
-
-		run = shared.OperationWrap(func() error {
-			if err := extractImage(hash, name, d); err != nil {
-				return err
-			}
-
-			if !c.isPrivileged() {
-				err = shiftRootfs(c, name, d)
-				if err != nil {
-					return err
-				}
-			}
-
-			err = templateApply(c, "create")
-			if err != nil {
-				return err
-			}
-
-			return nil
-		})
-	}
-
-	resources := make(map[string][]string)
-	resources["containers"] = []string{req.Name}
-
-	return &asyncResponse{run: run, resources: resources}
-}
-
-func createFromNone(d *Daemon, req *containerPostReq) Response {
-	args := DbCreateContainerArgs{
-		d:        d,
-		name:     req.Name,
-		ctype:    cTypeRegular,
-		config:   req.Config,
-		profiles: req.Profiles,
-		ephem:    req.Ephemeral,
-	}
-
-	_, err := dbCreateContainer(args)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	run := shared.OperationWrap(func() error {
-		c, err := newLxdContainer(req.Name, d)
-		if err != nil {
-			return err
-		}
-
-		err = templateApply(c, "create")
-		if err != nil {
-			return err
-		}
-
-		return nil
-	})
-
-	resources := make(map[string][]string)
-	resources["containers"] = []string{req.Name}
-
-	return &asyncResponse{run: run, resources: resources}
-}
-
-func extractShiftIfExists(d *Daemon, c *lxdContainer, hash string, name string) error {
-	if hash == "" {
-		return nil
-	}
-
-	_, err := dbImageGet(d.db, hash, false)
-	if err == nil {
-		if err := extractImage(hash, name, d); err != nil {
-			return err
-		}
-
-		if !c.isPrivileged() {
-			if err := shiftRootfs(c, name, d); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func createFromMigration(d *Daemon, req *containerPostReq) Response {
-
-	if req.Source.Mode != "pull" {
-		return NotImplemented
-	}
-
-	createArgs := DbCreateContainerArgs{
-		d:         d,
-		name:      req.Name,
-		ctype:     cTypeRegular,
-		config:    req.Config,
-		profiles:  req.Profiles,
-		ephem:     req.Ephemeral,
-		baseImage: req.Source.BaseImage,
-	}
-
-	_, err := dbCreateContainer(createArgs)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	c, err := newLxdContainer(req.Name, d)
-	if err != nil {
-		removeContainer(d, req.Name)
-		return SmartError(err)
-	}
-
-	// rsync complaisn if the parent directory for the rootfs sync doesn't
-	// exist
-	dpath := shared.VarPath("lxc", req.Name)
-	if err := os.MkdirAll(dpath, 0700); err != nil {
-		removeContainer(d, req.Name)
-		return InternalError(err)
-	}
-
-	if err := extractShiftIfExists(d, c, req.Source.BaseImage, req.Name); err != nil {
-		removeContainer(d, req.Name)
-		return InternalError(err)
-	}
-
-	config, err := shared.GetTLSConfig(d.certf, d.keyf)
-	if err != nil {
-		removeContainer(d, req.Name)
-		return InternalError(err)
-	}
-
-	args := migration.MigrationSinkArgs{
-		Url:       req.Source.Operation,
-		Dialer:    websocket.Dialer{TLSClientConfig: config},
-		Container: c.c,
-		Secrets:   req.Source.Websockets,
-		IdMapSet:  c.idmapset,
-	}
-
-	sink, err := migration.NewMigrationSink(&args)
-	if err != nil {
-		removeContainer(d, req.Name)
-		return BadRequest(err)
-	}
-
-	run := func() shared.OperationResult {
-		err := sink()
-		if err != nil {
-			removeContainer(d, req.Name)
-			return shared.OperationError(err)
-		}
-
-		c, err := newLxdContainer(req.Name, d)
-		if err != nil {
-			return shared.OperationError(err)
-		}
-
-		err = templateApply(c, "copy")
-		if err != nil {
-			return shared.OperationError(err)
-		}
-
-		return shared.OperationError(nil)
-	}
-
-	resources := make(map[string][]string)
-	resources["containers"] = []string{req.Name}
-
-	return &asyncResponse{run: run, resources: resources}
-}
-
-func createFromCopy(d *Daemon, req *containerPostReq) Response {
-	if req.Source.Source == "" {
-		return BadRequest(fmt.Errorf("must specify a source container"))
-	}
-
-	// Make sure the source exists.
-	source, err := newLxdContainer(req.Source.Source, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	if req.Config == nil {
-		config := make(map[string]string)
-		for key, value := range source.config {
-			if key[0:8] == "volatile" {
-				shared.Debugf("skipping: %s\n", key)
-				continue
-			}
-			req.Config[key] = value
-		}
-		req.Config = config
-	}
-
-	if req.Profiles == nil {
-		req.Profiles = source.profiles
-	}
-
-	args := DbCreateContainerArgs{
-		d:         d,
-		name:      req.Name,
-		ctype:     cTypeRegular,
-		config:    req.Config,
-		profiles:  req.Profiles,
-		ephem:     req.Ephemeral,
-		baseImage: req.Source.BaseImage,
-	}
-
-	_, err = dbCreateContainer(args)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	dpath := shared.VarPath("lxc", req.Name)
-	if err := os.MkdirAll(dpath, 0700); err != nil {
-		removeContainer(d, req.Name)
-		return InternalError(err)
-	}
-
-	if err := extractShiftIfExists(d, source, req.Source.BaseImage, req.Name); err != nil {
-		removeContainer(d, req.Name)
-		return InternalError(err)
-	}
-
-	var oldPath string
-	if shared.IsSnapshot(req.Source.Source) {
-		snappieces := strings.SplitN(req.Source.Source, "/", 2)
-		oldPath = migration.AddSlash(shared.VarPath("lxc",
-			snappieces[0],
-			"snapshots",
-			snappieces[1],
-			"rootfs"))
-	} else {
-		oldPath = migration.AddSlash(shared.VarPath("lxc", req.Source.Source, "rootfs"))
-	}
-	newPath := fmt.Sprintf("%s/%s", dpath, "rootfs")
-	run := func() shared.OperationResult {
-		output, err := exec.Command("rsync", "-a", "--devices", oldPath, newPath).CombinedOutput()
-
-		if err == nil && !source.isPrivileged() {
-			err = setUnprivUserAcl(source, dpath)
-			if err != nil {
-				shared.Debugf("Error adding acl for container root: falling back to chmod\n")
-				output, err := exec.Command("chmod", "+x", dpath).CombinedOutput()
-				if err != nil {
-					shared.Debugf("Error chmoding the container root\n")
-					shared.Debugf(string(output))
-					return shared.OperationError(err)
-				}
-			}
-		} else {
-			shared.Debugf("rsync failed:\n%s", output)
-			return shared.OperationError(err)
-		}
-
-		c, err := newLxdContainer(req.Name, d)
-		if err != nil {
-			return shared.OperationError(err)
-		}
-
-		err = templateApply(c, "copy")
-		if err != nil {
-			return shared.OperationError(err)
-		}
-
-		return shared.OperationError(nil)
-	}
-
-	resources := make(map[string][]string)
-	resources["containers"] = []string{req.Name, req.Source.Source}
-
-	return &asyncResponse{run: run, resources: resources}
-}
-
-func containersPost(d *Daemon, r *http.Request) Response {
-	shared.Debugf("responding to create")
-
-	if d.IdmapSet == nil {
-		return BadRequest(fmt.Errorf("shared's user has no subuids"))
-	}
-
-	req := containerPostReq{}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return BadRequest(err)
-	}
-
-	if req.Name == "" {
-		req.Name = strings.ToLower(petname.Generate(2, "-"))
-		shared.Debugf("no name provided, creating %s", req.Name)
-	}
-
-	switch req.Source.Type {
-	case "image":
-		return createFromImage(d, &req)
-	case "none":
-		return createFromNone(d, &req)
-	case "migration":
-		return createFromMigration(d, &req)
-	case "copy":
-		return createFromCopy(d, &req)
-	default:
-		return BadRequest(fmt.Errorf("unknown source type %s", req.Source.Type))
-	}
-
-}
-
-func removeContainerPath(d *Daemon, name string) error {
-	cpath := shared.VarPath("lxc", name)
-
-	backing_fs, err := shared.GetFilesystem(cpath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil
-		}
-
-		shared.Debugf("Error cleaning up %s: %s\n", cpath, err)
-		return err
-	}
-
-	vgname, vgnameIsSet, err := getServerConfigValue(d, "core.lvm_vg_name")
-	if err != nil {
-		return fmt.Errorf("Error checking server config: %v", err)
-	}
-
-	if vgnameIsSet {
-		output, err := exec.Command("umount", cpath).CombinedOutput()
-		if err != nil {
-			return fmt.Errorf("failed to unmount container path '%s'.\nError: %v\nOutput: %s", cpath, err, output)
-		}
-
-		err = shared.LVMRemoveLV(vgname, name)
-		if err != nil {
-			return fmt.Errorf("failed to remove deleted container LV: %v", err)
-		}
-
-	} else if backing_fs == "btrfs" {
-		exec.Command("btrfs", "subvolume", "delete", cpath).Run()
-	}
-
-	err = os.RemoveAll(cpath)
-	if err != nil {
-		shared.Debugf("Error cleaning up %s: %s\n", cpath, err)
-		return err
-	}
-
-	return nil
-}
-
-func removeContainer(d *Daemon, name string) error {
-	if err := containerDeleteSnapshots(d, name); err != nil {
-		return err
-	}
-
-	if err := removeContainerPath(d, name); err != nil {
-		return err
-	}
-
-	if err := dbRemoveContainer(d, name); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func btrfsCopyImage(hash string, name string, d *Daemon) error {
-	dpath := shared.VarPath("lxc", name)
-	imagefile := shared.VarPath("images", hash)
-	subvol := fmt.Sprintf("%s.btrfs", imagefile)
-
-	return exec.Command("btrfs", "subvolume", "snapshot", subvol, dpath).Run()
-}
-
-func extractImage(hash string, name string, d *Daemon) error {
-	/*
-	 * We want to use archive/tar for this, but that doesn't appear
-	 * to be working for us (see lxd/images.go)
-	 */
-	dpath := shared.VarPath("lxc", name)
-	imagefile := shared.VarPath("images", hash)
-
-	compression, _, err := detectCompression(imagefile)
-	if err != nil {
-		shared.Logf("Unkown compression type: %s", err)
-		removeContainer(d, name)
-		return err
-	}
-
-	args := []string{"-C", dpath, "--numeric-owner"}
-	switch compression {
-	case COMPRESSION_TAR:
-		args = append(args, "-xf")
-	case COMPRESSION_GZIP:
-		args = append(args, "-zxf")
-	case COMPRESSION_BZ2:
-		args = append(args, "--jxf")
-	case COMPRESSION_LZMA:
-		args = append(args, "--lzma", "-xf")
-	default:
-		args = append(args, "-Jxf")
-	}
-	args = append(args, imagefile)
-
-	output, err := exec.Command("tar", args...).Output()
-	if err != nil {
-		shared.Debugf("Untar of image: Output %s\nError %s\n", output, err)
-		removeContainer(d, name)
-		return err
-	}
-
-	return nil
-}
-
-func shiftRootfs(c *lxdContainer, name string, d *Daemon) error {
-	dpath := shared.VarPath("lxc", name)
-	rpath := shared.VarPath("lxc", name, "rootfs")
-	err := c.idmapset.ShiftRootfs(rpath)
-	if err != nil {
-		shared.Debugf("Shift of rootfs %s failed: %s\n", rpath, err)
-		removeContainer(d, name)
-		return err
-	}
-
-	/* Set an acl so the container root can descend the container dir */
-	err = setUnprivUserAcl(c, dpath)
-	if err != nil {
-		shared.Debugf("Error adding acl for container root: falling back to chmod\n")
-		output, err := exec.Command("chmod", "+x", dpath).CombinedOutput()
-		if err != nil {
-			shared.Debugf("Error chmoding the container root\n")
-			shared.Debugf(string(output))
-			return err
-		}
-	}
-
-	return nil
-}
-
-func setUnprivUserAcl(c *lxdContainer, dpath string) error {
-	if c.idmapset == nil {
-		return nil
-	}
-	uid, _ := c.idmapset.ShiftIntoNs(0, 0)
-	switch uid {
-	case -1:
-		shared.Debugf("setUnprivUserAcl: no root id mapping")
-		return nil
-	case 0:
-		return nil
-	}
-	acl := fmt.Sprintf("%d:rx", uid)
-	output, err := exec.Command("setfacl", "-m", acl, dpath).CombinedOutput()
-	if err != nil {
-		shared.Debugf("setfacl failed:\n%s", output)
-	}
-	return err
-}
-
-func dbRemoveContainer(d *Daemon, name string) error {
-	_, err := shared.DbExec(d.db, "DELETE FROM containers WHERE name=?", name)
-	return err
-}
-
-func dbGetContainerId(db *sql.DB, name string) (int, error) {
-	q := "SELECT id FROM containers WHERE name=?"
-	id := -1
-	arg1 := []interface{}{name}
-	arg2 := []interface{}{&id}
-	err := shared.DbQueryRowScan(db, q, arg1, arg2)
-	return id, err
-}
-
-type DbCreateContainerArgs struct {
-	d            *Daemon
-	name         string
-	ctype        containerType
-	config       map[string]string
-	profiles     []string
-	ephem        bool
-	baseImage    string
-	architecture int
-}
-
-func dbCreateContainer(args DbCreateContainerArgs) (int, error) {
-	id, err := dbGetContainerId(args.d.db, args.name)
-	if err == nil {
-		return 0, DbErrAlreadyDefined
-	}
-
-	if args.profiles == nil {
-		args.profiles = []string{"default"}
-	}
-
-	if args.baseImage != "" {
-		if args.config == nil {
-			args.config = map[string]string{}
-		}
-
-		args.config["volatile.baseImage"] = args.baseImage
-	}
-
-	tx, err := shared.DbBegin(args.d.db)
-	if err != nil {
-		return 0, err
-	}
-	ephem_int := 0
-	if args.ephem == true {
-		ephem_int = 1
-	}
-
-	str := fmt.Sprintf("INSERT INTO containers (name, architecture, type, ephemeral) VALUES (?, ?, ?, ?)")
-	stmt, err := tx.Prepare(str)
-	if err != nil {
-		tx.Rollback()
-		return 0, err
-	}
-	defer stmt.Close()
-	result, err := stmt.Exec(args.name, args.architecture, args.ctype, ephem_int)
-	if err != nil {
-		tx.Rollback()
-		return 0, err
-	}
-
-	id64, err := result.LastInsertId()
-	if err != nil {
-		tx.Rollback()
-		return 0, fmt.Errorf("Error inserting %s into database", args.name)
-	}
-	// TODO: is this really int64? we should fix it everywhere if so
-	id = int(id64)
-	if err := dbInsertContainerConfig(tx, id, args.config); err != nil {
-		tx.Rollback()
-		return 0, err
-	}
-
-	if err := dbInsertProfiles(tx, id, args.profiles); err != nil {
-		tx.Rollback()
-		return 0, err
-	}
-
-	return id, shared.TxCommit(tx)
-}
-
-var containersCmd = Command{name: "containers", get: containersGet, post: containersPost}
-
-func containerGet(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-	//cId, err := dbGetContainerId(d.db, name)  will need cId to get info
-	_, err := dbGetContainerId(d.db, name)
-	if err != nil {
-		return NotFound
-	}
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	state, err := c.RenderState()
-	if err != nil {
-		return InternalError(err)
-	}
-
-	targetPath := r.FormValue("log")
-	if strings.ToLower(targetPath) == "true" || targetPath == "1" {
-		fname := c.c.LogFile()
-
-		f, err := os.Open(fname)
-		if err != nil {
-			return InternalError(err)
-		}
-		defer f.Close()
-
-		log, err := shared.ReadLastNLines(f, 100)
-		if err != nil {
-			return InternalError(err)
-		}
-		state.Log = log
-	}
-
-	return SyncResponse(true, state)
-}
-
 func containerDeleteSnapshots(d *Daemon, cname string) error {
 	prefix := fmt.Sprintf("%s/", cname)
 	length := len(prefix)
@@ -1036,1773 +304,6 @@ func containerDeleteSnapshots(d *Daemon, cname string) error {
 
 	return nil
 }
-
-type containerConfigReq struct {
-	Profiles []string          `json:"profiles"`
-	Config   map[string]string `json:"config"`
-	Devices  shared.Devices    `json:"devices"`
-	Restore  string            `json:"restore"`
-}
-
-func containerSnapRestore(d *Daemon, name string, snap string) error {
-	// normalize snapshot name
-	if !shared.IsSnapshot(snap) {
-		snap = fmt.Sprintf("%s/%s", name, snap)
-	}
-
-	shared.Debugf("RESTORE => Restoring snapshot [%s] on container [%s]", snap, name)
-	/*
-	 * restore steps:
-	 * 1. stop container if already running
-	 * 2. overwrite existing config with snapshot config
-	 * 3. copy snapshot rootfs to container
-	 */
-	wasRunning := false
-	c, err := newLxdContainer(name, d)
-
-	if err != nil {
-		shared.Debugf("RESTORE => Error: newLxdContainer() failed for container", err)
-		return err
-	}
-
-	// 1. stop container
-	// TODO: stateful restore ?
-	if c.c.Running() {
-		wasRunning = true
-		if err = c.Stop(); err != nil {
-			shared.Debugf("RESTORE => Error: could not stop container", err)
-			return err
-		}
-		shared.Debugf("RESTORE => Stopped container %s", name)
-	}
-
-	// 2, replace config
-
-	// Make sure the source exists.
-	source, err := newLxdContainer(snap, d)
-	if err != nil {
-		shared.Debugf("RESTORE => Error: newLxdContainer() failed for snapshot", err)
-		return err
-	}
-
-	newConfig := containerConfigReq{}
-	newConfig.Config = source.config
-	newConfig.Profiles = source.profiles
-	newConfig.Devices = source.devices
-
-	err = containerReplaceConfig(d, c, name, newConfig)
-	if err != nil {
-		shared.Debugf("RESTORE => err #4", err)
-		return err
-	}
-
-	// 3. copy rootfs
-	// TODO: btrfs optimizations
-
-	containerRootPath := shared.VarPath("lxc", name)
-	// Check container dir already exists
-	fileinfo, err := os.Stat(path.Dir(containerRootPath))
-	if err != nil {
-		shared.Debugf("RESTORE => Could not stat %s to %s", containerRootPath, err)
-		return err
-	}
-
-	if !(fileinfo.IsDir()) {
-		shared.Debugf("RESTORE => containerRoot [%s] directory does not exist", containerRootPath)
-		return os.ErrNotExist
-	}
-
-	var snapshotRootFSPath string
-	snapshotRootFSPath = migration.AddSlash(snapshotRootfsDir(c, strings.SplitN(snap, "/", 2)[1]))
-
-	containerRootFSPath := migration.AddSlash(fmt.Sprintf("%s/%s", containerRootPath, "rootfs"))
-	shared.Debugf("RESTORE => Copying %s to %s", snapshotRootFSPath, containerRootFSPath)
-
-	rsync_verbosity := "-q"
-	if *debug {
-		rsync_verbosity = "-vi"
-	}
-
-	output, err := exec.Command("rsync", "-a", "-c", "-HAX", "--devices", "--delete", rsync_verbosity, snapshotRootFSPath, containerRootFSPath).CombinedOutput()
-	shared.Debugf("RESTORE => rsync output\n%s", output)
-
-	if err == nil && !source.isPrivileged() {
-		err = setUnprivUserAcl(c, containerRootPath)
-		if err != nil {
-			shared.Debugf("Error adding acl for container root: falling back to chmod\n")
-			output, err := exec.Command("chmod", "+x", containerRootPath).CombinedOutput()
-			if err != nil {
-				shared.Debugf("Error chmoding the container root\n")
-				shared.Debugf(string(output))
-				return err
-			}
-		}
-	} else {
-		shared.Debugf("rsync failed:\n%s", output)
-		return err
-	}
-
-	if wasRunning {
-		c.Start()
-	}
-
-	return nil
-}
-
-func dbClearContainerConfig(tx *sql.Tx, id int) error {
-	_, err := tx.Exec("DELETE FROM containers_config WHERE container_id=?", id)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec("DELETE FROM containers_profiles WHERE container_id=?", id)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec(`DELETE FROM containers_devices_config WHERE id IN
-		(SELECT containers_devices_config.id
-		 FROM containers_devices_config JOIN containers_devices
-		 ON containers_devices_config.container_device_id=containers_devices.id
-		 WHERE containers_devices.container_id=?)`, id)
-	if err != nil {
-		return err
-	}
-	_, err = tx.Exec("DELETE FROM containers_devices WHERE container_id=?", id)
-	return err
-}
-
-func dbInsertContainerConfig(tx *sql.Tx, id int, config map[string]string) error {
-	str := "INSERT INTO containers_config (container_id, key, value) values (?, ?, ?)"
-	stmt, err := tx.Prepare(str)
-	if err != nil {
-		return err
-	}
-	defer stmt.Close()
-
-	for k, v := range config {
-		if k == "raw.lxc" {
-			err := validateRawLxc(config["raw.lxc"])
-			if err != nil {
-				return err
-			}
-		}
-
-		if !ValidContainerConfigKey(k) {
-			return fmt.Errorf("Bad key: %s\n", k)
-		}
-
-		_, err = stmt.Exec(id, k, v)
-		if err != nil {
-			shared.Debugf("Error adding configuration item %s = %s to container %d\n",
-				k, v, id)
-			return err
-		}
-	}
-
-	return nil
-}
-
-func dbInsertProfiles(tx *sql.Tx, id int, profiles []string) error {
-	apply_order := 1
-	str := `INSERT INTO containers_profiles (container_id, profile_id, apply_order) VALUES
-		(?, (SELECT id FROM profiles WHERE name=?), ?);`
-	stmt, err := tx.Prepare(str)
-	if err != nil {
-		return err
-	}
-	defer stmt.Close()
-	for _, p := range profiles {
-		_, err = stmt.Exec(id, p, apply_order)
-		if err != nil {
-			shared.Debugf("Error adding profile %s to container: %s\n",
-				p, err)
-			return err
-		}
-		apply_order = apply_order + 1
-	}
-
-	return nil
-}
-
-// ExtractInterfaceFromConfigName returns "eth0" from "volatile.eth0.hwaddr",
-// or an error if the key does not match this pattern.
-func ExtractInterfaceFromConfigName(k string) (string, error) {
-
-	re := regexp.MustCompile("volatile\\.([^.]*)\\.hwaddr")
-	m := re.FindStringSubmatch(k)
-	if m != nil && len(m) > 1 {
-		return m[1], nil
-	}
-
-	return "", fmt.Errorf("%s did not match", k)
-}
-
-func ValidContainerConfigKey(k string) bool {
-	switch k {
-	case "limits.cpus":
-		return true
-	case "limits.memory":
-		return true
-	case "security.privileged":
-		return true
-	case "raw.apparmor":
-		return true
-	case "raw.lxc":
-		return true
-	case "volatile.baseImage":
-		return true
-	}
-
-	if _, err := ExtractInterfaceFromConfigName(k); err == nil {
-		return true
-	}
-
-	return strings.HasPrefix(k, "user.")
-}
-
-func emptyProfile(l []string) bool {
-	if len(l) == 0 {
-		return true
-	}
-	if len(l) == 1 && l[0] == "" {
-		return true
-	}
-	return false
-}
-
-/*
- * Update configuration, or, if 'restore:snapshot-name' is present, restore
- * the named snapshot
- */
-func containerPut(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return NotFound
-	}
-
-	configRaw := containerConfigReq{}
-	if err := json.NewDecoder(r.Body).Decode(&configRaw); err != nil {
-		return BadRequest(err)
-	}
-
-	var do = func() error { return nil }
-
-	if configRaw.Restore == "" {
-		// Update container configuration
-		do = func() error {
-			return containerReplaceConfig(d, c, name, configRaw)
-		}
-	} else {
-		// Snapshot Restore
-		do = func() error {
-			return containerSnapRestore(d, name, configRaw.Restore)
-		}
-	}
-
-	return AsyncResponse(shared.OperationWrap(do), nil)
-}
-
-func containerReplaceConfig(d *Daemon, ct *lxdContainer, name string, newConfig containerConfigReq) error {
-
-	/* check to see that the config actually applies to the container
-	 * successfully before saving it. in particular, raw.lxc and
-	 * raw.apparmor need to be parsed once to make sure they make sense.
-	 */
-	if err := ct.applyConfig(newConfig.Config, false); err != nil {
-		return err
-	}
-
-	tx, err := shared.DbBegin(d.db)
-	if err != nil {
-		return err
-	}
-
-	/* Update config or profiles */
-	if err = dbClearContainerConfig(tx, ct.id); err != nil {
-		shared.Debugf("Error clearing configuration for container %s\n", name)
-		tx.Rollback()
-		return err
-	}
-
-	if err = dbInsertContainerConfig(tx, ct.id, newConfig.Config); err != nil {
-		shared.Debugf("Error inserting configuration for container %s\n", name)
-		tx.Rollback()
-		return err
-	}
-
-	/* handle profiles */
-	if emptyProfile(newConfig.Profiles) {
-		_, err := tx.Exec("DELETE from containers_profiles where container_id=?", ct.id)
-		if err != nil {
-			tx.Rollback()
-			return err
-		}
-	} else {
-		if err := dbInsertProfiles(tx, ct.id, newConfig.Profiles); err != nil {
-
-			tx.Rollback()
-			return err
-		}
-	}
-
-	err = AddDevices(tx, "container", ct.id, newConfig.Devices)
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-
-	return shared.TxCommit(tx)
-}
-
-type containerPostBody struct {
-	Migration bool   `json:"migration"`
-	Name      string `json:"name"`
-}
-
-func containerPost(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	buf, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return InternalError(err)
-	}
-
-	body := containerPostBody{}
-	if err := json.Unmarshal(buf, &body); err != nil {
-		return BadRequest(err)
-	}
-
-	if body.Migration {
-		ws, err := migration.NewMigrationSource(c.c)
-		if err != nil {
-			return InternalError(err)
-		}
-
-		return AsyncResponseWithWs(ws, nil)
-	} else {
-		if c.c.Running() {
-			return BadRequest(fmt.Errorf("renaming of running container not allowed"))
-		}
-
-		args := DbCreateContainerArgs{
-			d:            d,
-			name:         body.Name,
-			ctype:        cTypeRegular,
-			config:       c.config,
-			profiles:     c.profiles,
-			ephem:        c.ephemeral,
-			baseImage:    c.config["volatile.baseImage"],
-			architecture: c.architecture,
-		}
-
-		_, err := dbCreateContainer(args)
-		if err != nil {
-			return SmartError(err)
-		}
-
-		run := func() error {
-			oldPath := fmt.Sprintf("%s/", shared.VarPath("lxc", c.name))
-			newPath := fmt.Sprintf("%s/", shared.VarPath("lxc", body.Name))
-
-			if err := os.Rename(oldPath, newPath); err != nil {
-				return err
-			}
-
-			removeContainer(d, c.name)
-			return nil
-		}
-
-		return AsyncResponse(shared.OperationWrap(run), nil)
-	}
-}
-
-func containerDelete(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-	_, err := dbGetContainerId(d.db, name)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	rmct := func() error {
-		return removeContainer(d, name)
-	}
-
-	return AsyncResponse(shared.OperationWrap(rmct), nil)
-}
-
-var containerCmd = Command{name: "containers/{name}", get: containerGet, put: containerPut, delete: containerDelete, post: containerPost}
-
-func containerStateGet(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	state, err := c.RenderState()
-	if err != nil {
-		return InternalError(err)
-	}
-
-	return SyncResponse(true, state.Status)
-}
-
-type containerStatePutReq struct {
-	Action  string `json:"action"`
-	Timeout int    `json:"timeout"`
-	Force   bool   `json:"force"`
-}
-
-type lxdContainer struct {
-	c            *lxc.Container
-	daemon       *Daemon
-	id           int
-	name         string
-	config       map[string]string
-	profiles     []string
-	devices      shared.Devices
-	architecture int
-	ephemeral    bool
-	idmapset     *shared.IdmapSet
-}
-
-func getIps(c *lxc.Container) []shared.Ip {
-	ips := []shared.Ip{}
-	names, err := c.Interfaces()
-	if err != nil {
-		return ips
-	}
-	for _, n := range names {
-		addresses, err := c.IPAddress(n)
-		if err != nil {
-			continue
-		}
-
-		veth := ""
-
-		for i := 0; i < len(c.ConfigItem("lxc.network")); i++ {
-			nicName := c.RunningConfigItem(fmt.Sprintf("lxc.network.%d.name", i))[0]
-			if nicName != n {
-				continue
-			}
-
-			interfaceType := c.RunningConfigItem(fmt.Sprintf("lxc.network.%d.type", i))
-			if interfaceType[0] != "veth" {
-				continue
-			}
-
-			veth = c.RunningConfigItem(fmt.Sprintf("lxc.network.%d.veth.pair", i))[0]
-			break
-		}
-
-		for _, a := range addresses {
-			ip := shared.Ip{Interface: n, Address: a, HostVeth: veth}
-			if net.ParseIP(a).To4() == nil {
-				ip.Protocol = "IPV6"
-			} else {
-				ip.Protocol = "IPV4"
-			}
-			ips = append(ips, ip)
-		}
-	}
-	return ips
-}
-
-func NewStatus(c *lxc.Container, state lxc.State) shared.ContainerStatus {
-	status := shared.ContainerStatus{State: state.String(), StateCode: shared.State(int(state))}
-	if state == lxc.RUNNING {
-		status.Init = c.InitPid()
-		status.Ips = getIps(c)
-	}
-	return status
-}
-
-func (c *lxdContainer) RenderState() (*shared.ContainerState, error) {
-	devices, err := dbGetDevices(c.daemon.db, c.name, false)
-	if err != nil {
-		return nil, err
-	}
-
-	config, err := dbGetConfig(c.daemon.db, c.id)
-	if err != nil {
-		return nil, err
-	}
-
-	return &shared.ContainerState{
-		Name:            c.name,
-		Profiles:        c.profiles,
-		Config:          config,
-		ExpandedConfig:  c.config,
-		Userdata:        []byte{},
-		Status:          NewStatus(c.c, c.c.State()),
-		Devices:         devices,
-		ExpandedDevices: c.devices,
-		Ephemeral:       c.ephemeral,
-	}, nil
-}
-
-func (c *lxdContainer) Start() error {
-
-	f, err := ioutil.TempFile("", "lxd_lxc_startconfig_")
-	if err != nil {
-		return err
-	}
-	configPath := f.Name()
-	if err = f.Chmod(0600); err != nil {
-		f.Close()
-		os.Remove(configPath)
-		return err
-	}
-	f.Close()
-
-	err = c.c.SaveConfigFile(configPath)
-	if err != nil {
-		return err
-	}
-
-	err = templateApply(c, "start")
-	if err != nil {
-		return err
-	}
-
-	cmd := exec.Command(os.Args[0], "forkstart", c.name, c.daemon.lxcpath, configPath)
-	err = cmd.Run()
-
-	if err != nil {
-		err = fmt.Errorf("Error calling 'lxd forkstart': %v", err)
-	}
-
-	if err == nil && c.ephemeral == true {
-		containerWatchEphemeral(c)
-	}
-
-	return err
-}
-
-func (c *lxdContainer) Reboot() error {
-	return c.c.Reboot()
-}
-
-func (c *lxdContainer) Freeze() error {
-	return c.c.Freeze()
-}
-
-func (c *lxdContainer) isPrivileged() bool {
-	switch strings.ToLower(c.config["security.privileged"]) {
-	case "1":
-		return true
-	case "true":
-		return true
-	}
-	return false
-}
-
-func (c *lxdContainer) Shutdown(timeout time.Duration) error {
-	return c.c.Shutdown(timeout)
-}
-
-func (c *lxdContainer) Stop() error {
-	return c.c.Stop()
-}
-
-func (c *lxdContainer) Unfreeze() error {
-	return c.c.Unfreeze()
-}
-
-func validateRawLxc(rawLxc string) error {
-	for _, line := range strings.Split(rawLxc, "\n") {
-		membs := strings.SplitN(line, "=", 2)
-		if strings.ToLower(strings.Trim(membs[0], " \t")) == "lxc.logfile" {
-			return fmt.Errorf("setting lxc.logfile is not allowed")
-		}
-	}
-
-	return nil
-}
-
-func (d *lxdContainer) applyConfig(config map[string]string, fromProfile bool) error {
-	var err error
-	for k, v := range config {
-		switch k {
-		case "limits.cpus":
-			// TODO - Come up with a way to choose cpus for multiple
-			// containers
-			var vint int
-			count, err := fmt.Sscanf(v, "%d", &vint)
-			if err != nil {
-				return err
-			}
-			if count != 1 || vint < 0 || vint > 65000 {
-				return fmt.Errorf("Bad cpu limit: %s\n", v)
-			}
-			cpuset := fmt.Sprintf("0-%d", vint-1)
-			err = d.c.SetConfigItem("lxc.cgroup.cpuset.cpus", cpuset)
-		case "limits.memory":
-			err = d.c.SetConfigItem("lxc.cgroup.memory.limit_in_bytes", v)
-
-		default:
-			if strings.HasPrefix(k, "user.") {
-				// ignore for now
-				err = nil
-			}
-
-			/* Things like security.privileged need to be propagated */
-			d.config[k] = v
-		}
-		if err != nil {
-			shared.Debugf("error setting %s: %q\n", k, err)
-			return err
-		}
-	}
-
-	if fromProfile {
-		return nil
-	}
-
-	if lxcConfig, ok := config["raw.lxc"]; ok {
-		if err := validateRawLxc(lxcConfig); err != nil {
-			return err
-		}
-
-		f, err := ioutil.TempFile("", "lxd_config_")
-		if err != nil {
-			return err
-		}
-
-		err = shared.WriteAll(f, []byte(lxcConfig))
-		f.Close()
-		defer os.Remove(f.Name())
-		if err != nil {
-			return err
-		}
-
-		if err := d.c.LoadConfigFile(f.Name()); err != nil {
-			return fmt.Errorf("problem applying raw.lxc, perhaps there is a syntax error?")
-		}
-	}
-
-	return nil
-}
-
-func applyProfile(daemon *Daemon, d *lxdContainer, p string) error {
-	q := `SELECT key, value FROM profiles_config
-		JOIN profiles ON profiles.id=profiles_config.profile_id
-		WHERE profiles.name=?`
-	var k, v string
-	inargs := []interface{}{p}
-	outfmt := []interface{}{k, v}
-	result, err := shared.DbQueryScan(daemon.db, q, inargs, outfmt)
-
-	if err != nil {
-		return err
-	}
-
-	config := map[string]string{}
-	for _, r := range result {
-		k = r[0].(string)
-		v = r[1].(string)
-
-		shared.Debugf("applying %s: %s", k, v)
-		if k == "raw.lxc" {
-			if _, ok := d.config["raw.lxc"]; ok {
-				shared.Debugf("ignoring overridden raw.lxc from profile")
-				continue
-			}
-		}
-
-		config[k] = v
-	}
-
-	newdevs, err := dbGetDevices(daemon.db, p, true)
-	if err != nil {
-		return err
-	}
-	for k, v := range newdevs {
-		d.devices[k] = v
-	}
-
-	return d.applyConfig(config, true)
-}
-
-// GenerateMacAddr generates a mac address from a string template:
-// e.g. "00:11:22:xx:xx:xx" -> "00:11:22:af:3e:51"
-func GenerateMacAddr(template string) (string, error) {
-	ret := bytes.Buffer{}
-
-	for _, c := range template {
-		if c == 'x' {
-			c, err := rand.Int(rand.Reader, big.NewInt(16))
-			if err != nil {
-				return "", err
-			}
-			ret.WriteString(fmt.Sprintf("%x", c.Int64()))
-		} else {
-			ret.WriteString(string(c))
-		}
-	}
-
-	return ret.String(), nil
-}
-
-func (c *lxdContainer) updateContainerHWAddr(k, v string) {
-	for name, d := range c.devices {
-		if d["type"] != "nic" {
-			continue
-		}
-
-		for key, _ := range c.config {
-			device, err := ExtractInterfaceFromConfigName(key)
-			if err == nil && device == name {
-				d["hwaddr"] = v
-				c.config[key] = v
-				return
-			}
-		}
-	}
-}
-
-func (c *lxdContainer) setupMacAddresses(d *Daemon) error {
-	newConfigEntries := map[string]string{}
-
-	for name, d := range c.devices {
-		if d["type"] != "nic" {
-			continue
-		}
-
-		found := false
-
-		for key, val := range c.config {
-			device, err := ExtractInterfaceFromConfigName(key)
-			if err == nil && device == name {
-				found = true
-				d["hwaddr"] = val
-			}
-		}
-
-		if !found {
-			var hwaddr string
-			var err error
-			if d["hwaddr"] != "" {
-				hwaddr, err = GenerateMacAddr(d["hwaddr"])
-				if err != nil {
-					return err
-				}
-			} else {
-				hwaddr, err = GenerateMacAddr("00:16:3e:xx:xx:xx")
-				if err != nil {
-					return err
-				}
-			}
-
-			if hwaddr != d["hwaddr"] {
-				d["hwaddr"] = hwaddr
-				key := fmt.Sprintf("volatile.%s.hwaddr", name)
-				c.config[key] = hwaddr
-				newConfigEntries[key] = hwaddr
-			}
-		}
-	}
-
-	if len(newConfigEntries) > 0 {
-
-		tx, err := shared.DbBegin(d.db)
-		if err != nil {
-			return err
-		}
-
-		/*
-		 * My logic may be flawed here, but it seems to me that one of
-		 * the following must be true:
-		 * 1. The current database entry equals what we had stored.
-		 *    Our update akes precedence
-		 * 2. The current database entry is different from what we had
-		 *    stored.  Someone updated it since we last grabbed the
-		 *    container configuration.  So either
-		 *    a. it contains 'x' and is a template.  We have generated
-		 *       a real mac, so our update takes precedence
-		 *    b. it contains no 'x' and is an hwaddr, not template.  We
-		 *       defer to the racer's update since it may be actually
-		 *       starting the container.
-		 */
-		str := "INSERT INTO containers_config (container_id, key, value) values (?, ?, ?)"
-		stmt, err := tx.Prepare(str)
-		if err != nil {
-			tx.Rollback()
-			return err
-		}
-		defer stmt.Close()
-
-		ustr := "UPDATE containers_config SET value=? WHERE container_id=? AND key=?"
-		ustmt, err := tx.Prepare(ustr)
-		if err != nil {
-			tx.Rollback()
-			return err
-		}
-		defer ustmt.Close()
-
-		qstr := "SELECT value FROM containers_config WHERE container_id=? AND key=?"
-		qstmt, err := tx.Prepare(qstr)
-		if err != nil {
-			tx.Rollback()
-			return err
-		}
-		defer qstmt.Close()
-
-		for k, v := range newConfigEntries {
-			var racer string
-			err := qstmt.QueryRow(c.id, k).Scan(&racer)
-			if err == sql.ErrNoRows {
-				_, err = stmt.Exec(c.id, k, v)
-				if err != nil {
-					shared.Debugf("Error adding mac address to container\n")
-					tx.Rollback()
-					return err
-				}
-			} else if err != nil {
-				tx.Rollback()
-				return err
-			} else if strings.Contains(racer, "x") {
-				_, err = ustmt.Exec(v, c.id, k)
-				if err != nil {
-					shared.Debugf("Error updating mac address to container\n")
-					tx.Rollback()
-					return err
-				}
-			} else {
-				// we accept the racing task's update
-				c.updateContainerHWAddr(k, v)
-			}
-		}
-
-		err = shared.TxCommit(tx)
-		if err != nil {
-			fmt.Printf("setupMacAddresses: (TxCommit) error %s\n", err)
-		}
-		return err
-	}
-
-	return nil
-}
-
-func (c *lxdContainer) applyIdmapSet() error {
-	if c.idmapset == nil {
-		return nil
-	}
-	lines := c.idmapset.ToLxcString()
-	for _, line := range lines {
-		err := c.c.SetConfigItem("lxc.id_map", line)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *lxdContainer) applyDevices() error {
-	var keys []string
-	for k := range c.devices {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	for _, name := range keys {
-		d := c.devices[name]
-		if name == "type" {
-			continue
-		}
-
-		configs, err := DeviceToLxc(d)
-		if err != nil {
-			return fmt.Errorf("Failed configuring device %s: %s\n", name, err)
-		}
-		for _, line := range configs {
-			err := c.c.SetConfigItem(line[0], line[1])
-			if err != nil {
-				return fmt.Errorf("Failed configuring device %s: %s\n", name, err)
-			}
-		}
-	}
-	return nil
-}
-
-func newLxdContainer(name string, daemon *Daemon) (*lxdContainer, error) {
-	d := &lxdContainer{}
-	d.daemon = daemon
-	ephem_int := -1
-	d.ephemeral = false
-	d.architecture = -1
-	d.id = -1
-
-	templateConfBase := "ubuntu"
-	templateConfDir := os.Getenv("LXC_TEMPLATE_CONFIG")
-	if templateConfDir == "" {
-		templateConfDir = "/usr/share/lxc/config"
-	}
-
-	q := "SELECT id, architecture, ephemeral FROM containers WHERE name=?"
-	arg1 := []interface{}{name}
-	arg2 := []interface{}{&d.id, &d.architecture, &ephem_int}
-	err := shared.DbQueryRowScan(daemon.db, q, arg1, arg2)
-	if err != nil {
-		return nil, err
-	}
-	if d.id == -1 {
-		return nil, fmt.Errorf("Unknown container")
-	}
-
-	if ephem_int == 1 {
-		d.ephemeral = true
-	}
-
-	c, err := lxc.NewContainer(name, daemon.lxcpath)
-	if err != nil {
-		return nil, err
-	}
-	d.c = c
-
-	dir := shared.LogPath(c.Name())
-	err = os.MkdirAll(dir, 0700)
-	if err != nil {
-		return nil, err
-	}
-
-	if err = d.c.SetLogFile(filepath.Join(dir, "lxc.log")); err != nil {
-		return nil, err
-	}
-
-	personality, err := shared.ArchitecturePersonality(d.architecture)
-	if err == nil {
-		err = c.SetConfigItem("lxc.arch", personality)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	err = c.SetConfigItem("lxc.include", fmt.Sprintf("%s/%s.common.conf", templateConfDir, templateConfBase))
-	if err != nil {
-		return nil, err
-	}
-
-	if !d.isPrivileged() {
-		err = c.SetConfigItem("lxc.include", fmt.Sprintf("%s/%s.userns.conf", templateConfDir, templateConfBase))
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	config, err := dbGetConfig(daemon.db, d.id)
-	if err != nil {
-		return nil, err
-	}
-	d.config = config
-
-	profiles, err := dbGetProfiles(daemon.db, d.id)
-	if err != nil {
-		return nil, err
-	}
-	d.profiles = profiles
-	d.devices = shared.Devices{}
-	d.name = name
-
-	rootfsPath := shared.VarPath("lxc", name, "rootfs")
-	err = c.SetConfigItem("lxc.rootfs", rootfsPath)
-	if err != nil {
-		return nil, err
-	}
-	err = c.SetConfigItem("lxc.loglevel", "0")
-	if err != nil {
-		return nil, err
-	}
-	err = c.SetConfigItem("lxc.utsname", name)
-	if err != nil {
-		return nil, err
-	}
-	err = c.SetConfigItem("lxc.tty", "0")
-	if err != nil {
-		return nil, err
-	}
-
-	if err := setupDevLxdMount(c); err != nil {
-		return nil, err
-	}
-
-	/* apply profiles */
-	for _, p := range profiles {
-		err := applyProfile(daemon, d, p)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	/* get container_devices */
-	newdevs, err := dbGetDevices(daemon.db, d.name, false)
-	if err != nil {
-		return nil, err
-	}
-
-	for k, v := range newdevs {
-		d.devices[k] = v
-	}
-
-	if err := d.setupMacAddresses(daemon); err != nil {
-		return nil, err
-	}
-
-	/* now add the lxc.* entries for the configured devices */
-	err = d.applyDevices()
-	if err != nil {
-		return nil, err
-	}
-
-	if !d.isPrivileged() {
-		d.idmapset = daemon.IdmapSet // TODO - per-tenant idmaps
-	}
-
-	err = d.applyIdmapSet()
-	if err != nil {
-		return nil, err
-	}
-
-	err = d.applyConfig(d.config, false)
-	if err != nil {
-		return nil, err
-	}
-
-	return d, nil
-}
-
-func containerStatePut(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-
-	raw := containerStatePutReq{}
-
-	// We default to -1 (i.e. no timeout) here instead of 0 (instant
-	// timeout).
-	raw.Timeout = -1
-
-	if err := json.NewDecoder(r.Body).Decode(&raw); err != nil {
-		return BadRequest(err)
-	}
-
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	var do func() error
-	switch shared.ContainerAction(raw.Action) {
-	case shared.Start:
-		do = c.Start
-	case shared.Stop:
-		if raw.Timeout == 0 || raw.Force {
-			do = c.Stop
-		} else {
-			do = func() error { return c.Shutdown(time.Duration(raw.Timeout) * time.Second) }
-		}
-	case shared.Restart:
-		do = c.Reboot
-	case shared.Freeze:
-		do = c.Freeze
-	case shared.Unfreeze:
-		do = c.Unfreeze
-	default:
-		return BadRequest(fmt.Errorf("unknown action %s", raw.Action))
-	}
-
-	return AsyncResponse(shared.OperationWrap(do), nil)
-}
-
-var containerStateCmd = Command{name: "containers/{name}/state", get: containerStateGet, put: containerStatePut}
-
-func containerFileHandler(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	targetPath := r.FormValue("path")
-	if targetPath == "" {
-		return BadRequest(fmt.Errorf("missing path argument"))
-	}
-
-	if !c.c.Running() {
-		return BadRequest(fmt.Errorf("container is not running"))
-	}
-
-	switch r.Method {
-	case "GET":
-		return containerFileGet(c.c.InitPid(), r, targetPath)
-	case "POST":
-		return containerFilePut(c.c.InitPid(), r, targetPath, c.idmapset)
-	default:
-		return NotFound
-	}
-}
-
-func containerFileGet(pid int, r *http.Request, path string) Response {
-
-	/*
-	 * Copy out of the ns to a temporary file, and then use that to serve
-	 * the request from. This prevents us from having to worry about stuff
-	 * like people breaking out of the container root by symlinks or
-	 * ../../../s etc. in the path, since we can just rely on the kernel
-	 * for correctness.
-	 */
-	temp, err := ioutil.TempFile("", "lxd_forkgetfile_")
-	if err != nil {
-		return InternalError(err)
-	}
-	defer temp.Close()
-
-	cmd := exec.Command(
-		os.Args[0],
-		"forkgetfile",
-		temp.Name(),
-		fmt.Sprintf("%d", pid),
-		path,
-	)
-	if out, err := cmd.CombinedOutput(); err != nil {
-		return InternalError(fmt.Errorf("%s: %s", err.Error(), string(out)))
-	}
-
-	fi, err := temp.Stat()
-	if err != nil {
-		return SmartError(err)
-	}
-
-	/*
-	 * Unfortunately, there's no portable way to do this:
-	 * https://groups.google.com/forum/#!topic/golang-nuts/tGYjYyrwsGM
-	 * https://groups.google.com/forum/#!topic/golang-nuts/ywS7xQYJkHY
-	 */
-	sb := fi.Sys().(*syscall.Stat_t)
-	headers := map[string]string{
-		"X-LXD-uid":  strconv.FormatUint(uint64(sb.Uid), 10),
-		"X-LXD-gid":  strconv.FormatUint(uint64(sb.Gid), 10),
-		"X-LXD-mode": fmt.Sprintf("%04o", fi.Mode()&os.ModePerm),
-	}
-
-	return FileResponse(r, temp.Name(), filepath.Base(path), headers, true)
-}
-
-func containerFilePut(pid int, r *http.Request, p string, idmapset *shared.IdmapSet) Response {
-
-	uid, gid, mode, err := shared.ParseLXDFileHeaders(r.Header)
-	if err != nil {
-		return BadRequest(err)
-	}
-	uid, gid = idmapset.ShiftIntoNs(uid, gid)
-	if uid == -1 || gid == -1 {
-		return BadRequest(fmt.Errorf("unmapped uid or gid specified"))
-	}
-
-	temp, err := ioutil.TempFile("", "lxd_forkputfile_")
-	if err != nil {
-		return InternalError(err)
-	}
-	defer func() {
-		temp.Close()
-		os.Remove(temp.Name())
-	}()
-
-	_, err = io.Copy(temp, r.Body)
-	if err != nil {
-		return InternalError(err)
-	}
-
-	cmd := exec.Command(
-		os.Args[0],
-		"forkputfile",
-		temp.Name(),
-		fmt.Sprintf("%d", pid),
-		p,
-		fmt.Sprintf("%d", uid),
-		fmt.Sprintf("%d", gid),
-		fmt.Sprintf("%d", mode&os.ModePerm),
-	)
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return InternalError(fmt.Errorf("%s: %s", err.Error(), string(out)))
-	}
-
-	return EmptySyncResponse
-}
-
-var containerFileCmd = Command{name: "containers/{name}/files", get: containerFileHandler, post: containerFileHandler}
-
-func snapshotsDir(c *lxdContainer) string {
-	return shared.VarPath("lxc", c.name, "snapshots")
-}
-
-func snapshotDir(c *lxdContainer, name string) string {
-	return path.Join(snapshotsDir(c), name)
-}
-
-func snapshotStateDir(c *lxdContainer, name string) string {
-	return path.Join(snapshotDir(c, name), "state")
-}
-
-func snapshotRootfsDir(c *lxdContainer, name string) string {
-	return path.Join(snapshotDir(c, name), "rootfs")
-}
-
-func containerSnapshotsGet(d *Daemon, r *http.Request) Response {
-	recursion_str := r.FormValue("recursion")
-	recursion, err := strconv.Atoi(recursion_str)
-	if err != nil {
-		recursion = 0
-	}
-
-	cname := mux.Vars(r)["name"]
-	c, err := newLxdContainer(cname, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	regexp := fmt.Sprintf("%s/", cname)
-	length := len(regexp)
-	q := "SELECT name FROM containers WHERE type=? AND SUBSTR(name,1,?)=?"
-	var name string
-	inargs := []interface{}{cTypeSnapshot, length, regexp}
-	outfmt := []interface{}{name}
-	results, err := shared.DbQueryScan(d.db, q, inargs, outfmt)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	var result_string []string
-	var result_map []shared.Jmap
-
-	for _, r := range results {
-		name = r[0].(string)
-		if recursion == 0 {
-			url := fmt.Sprintf("/%s/containers/%s/snapshots/%s", shared.APIVersion, cname, name)
-			result_string = append(result_string, url)
-		} else {
-			_, err := os.Stat(snapshotStateDir(c, name))
-			body := shared.Jmap{"name": name, "stateful": err == nil}
-			result_map = append(result_map, body)
-		}
-
-	}
-
-	if recursion == 0 {
-		return SyncResponse(true, result_string)
-	} else {
-		return SyncResponse(true, result_map)
-	}
-
-}
-
-/*
- * Note, the code below doesn't deal with snapshots of snapshots.
- * To do that, we'll need to weed out based on # slashes in names
- */
-func nextSnapshot(d *Daemon, name string) int {
-	base := fmt.Sprintf("%s/snap", name)
-	length := len(base)
-	q := fmt.Sprintf("SELECT MAX(name) FROM containers WHERE type=? AND SUBSTR(name,1,?)=?")
-	var numstr string
-	inargs := []interface{}{cTypeSnapshot, length, base}
-	outfmt := []interface{}{numstr}
-	results, err := shared.DbQueryScan(d.db, q, inargs, outfmt)
-	if err != nil {
-		return 0
-	}
-	max := 0
-
-	for _, r := range results {
-		numstr = r[0].(string)
-		if len(numstr) <= length {
-			continue
-		}
-		substr := numstr[length:]
-		var num int
-		count, err := fmt.Sscanf(substr, "%d", &num)
-		if err != nil || count != 1 {
-			continue
-		}
-		if num >= max {
-			max = num + 1
-		}
-	}
-
-	return max
-}
-
-func containerSnapshotsPost(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-
-	/*
-	 * snapshot is a three step operation:
-	 * 1. choose a new name
-	 * 2. copy the database info over
-	 * 3. copy over the rootfs
-	 */
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	raw := shared.Jmap{}
-	if err := json.NewDecoder(r.Body).Decode(&raw); err != nil {
-		return BadRequest(err)
-	}
-
-	snapshotName, err := raw.GetString("name")
-	if err != nil || snapshotName == "" {
-		// come up with a name
-		i := nextSnapshot(d, name)
-		snapshotName = fmt.Sprintf("snap%d", i)
-	}
-
-	stateful, err := raw.GetBool("stateful")
-	if err != nil {
-		return BadRequest(err)
-	}
-
-	fullName := fmt.Sprintf("%s/%s", name, snapshotName)
-	snapDir := snapshotDir(c, snapshotName)
-	if shared.PathExists(snapDir) {
-		return Conflict
-	}
-
-	err = os.MkdirAll(snapDir, 0700)
-	if err != nil {
-		return InternalError(err)
-	}
-
-	snapshot := func() error {
-
-		StateDir := snapshotStateDir(c, snapshotName)
-		err = os.MkdirAll(StateDir, 0700)
-		if err != nil {
-			return err
-		}
-
-		if stateful {
-			// TODO - shouldn't we freeze for the duration of rootfs snapshot below?
-			if !c.c.Running() {
-				return fmt.Errorf("Container not running\n")
-			}
-			opts := lxc.CheckpointOptions{Directory: StateDir, Stop: true, Verbose: true}
-			if err := c.c.Checkpoint(opts); err != nil {
-				return err
-			}
-		}
-
-		/* Create the db info */
-		args := DbCreateContainerArgs{
-			d:            d,
-			name:         fullName,
-			ctype:        cTypeSnapshot,
-			config:       c.config,
-			profiles:     c.profiles,
-			ephem:        c.ephemeral,
-			baseImage:    c.config["volatile.baseImage"],
-			architecture: c.architecture,
-		}
-
-		_, err := dbCreateContainer(args)
-		if err != nil {
-			return err
-		}
-
-		/* Create the directory and rootfs, set perms */
-		/* Copy the rootfs */
-		oldPath := fmt.Sprintf("%s/", shared.VarPath("lxc", name, "rootfs"))
-		newPath := snapshotRootfsDir(c, snapshotName)
-		err = exec.Command("rsync", "-a", "--devices", oldPath, newPath).Run()
-		return err
-	}
-
-	return AsyncResponse(shared.OperationWrap(snapshot), nil)
-}
-
-var containerSnapshotsCmd = Command{name: "containers/{name}/snapshots", get: containerSnapshotsGet, post: containerSnapshotsPost}
-
-func dbRemoveSnapshot(d *Daemon, cname string, sname string) {
-	name := fmt.Sprintf("%s/%s", cname, sname)
-	_, _ = shared.DbExec(d.db, "DELETE FROM containers WHERE type=? AND name=?", cTypeSnapshot, name)
-}
-
-func snapshotHandler(d *Daemon, r *http.Request) Response {
-	containerName := mux.Vars(r)["name"]
-	c, err := newLxdContainer(containerName, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	snapshotName := mux.Vars(r)["snapshotName"]
-	dir := snapshotDir(c, snapshotName)
-
-	_, err = os.Stat(dir)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	switch r.Method {
-	case "GET":
-		return snapshotGet(c, snapshotName)
-	case "POST":
-		return snapshotPost(r, c, snapshotName)
-	case "DELETE":
-		return snapshotDelete(d, c, snapshotName)
-	default:
-		return NotFound
-	}
-}
-
-func snapshotGet(c *lxdContainer, name string) Response {
-	_, err := os.Stat(snapshotStateDir(c, name))
-	body := shared.Jmap{"name": name, "stateful": err == nil}
-	return SyncResponse(true, body)
-}
-
-func snapshotPost(r *http.Request, c *lxdContainer, oldName string) Response {
-	raw := shared.Jmap{}
-	if err := json.NewDecoder(r.Body).Decode(&raw); err != nil {
-		return BadRequest(err)
-	}
-
-	newName, err := raw.GetString("name")
-	if err != nil {
-		return BadRequest(err)
-	}
-
-	oldDir := snapshotDir(c, oldName)
-	newDir := snapshotDir(c, newName)
-
-	_, err = os.Stat(newDir)
-	if !os.IsNotExist(err) {
-		return InternalError(err)
-	} else if err == nil {
-		return Conflict
-	}
-
-	/*
-	 * TODO: do we need to do something more intelligent here? We probably
-	 * shouldn't do anything for stateful snapshots, since changing the fs
-	 * out from under criu will cause it to fail, but it may be useful to
-	 * do something for stateless ones.
-	 */
-	rename := func() error { return os.Rename(oldDir, newDir) }
-	return AsyncResponse(shared.OperationWrap(rename), nil)
-}
-
-func snapshotDelete(d *Daemon, c *lxdContainer, name string) Response {
-	dbRemoveSnapshot(d, c.name, name)
-	dir := snapshotDir(c, name)
-	remove := func() error { return os.RemoveAll(dir) }
-	return AsyncResponse(shared.OperationWrap(remove), nil)
-}
-
-var containerSnapshotCmd = Command{name: "containers/{name}/snapshots/{snapshotName}", get: snapshotHandler, post: snapshotHandler, delete: snapshotHandler}
-
-type execWs struct {
-	command          []string
-	container        *lxc.Container
-	rootUid          int
-	rootGid          int
-	options          lxc.AttachOptions
-	conns            map[int]*websocket.Conn
-	allConnected     chan bool
-	controlConnected chan bool
-	interactive      bool
-	done             chan shared.OperationResult
-	fds              map[int]string
-}
-
-func (s *execWs) Metadata() interface{} {
-	fds := shared.Jmap{}
-	for fd, secret := range s.fds {
-		if fd == -1 {
-			fds["control"] = secret
-		} else {
-			fds[strconv.Itoa(fd)] = secret
-		}
-	}
-
-	return shared.Jmap{"fds": fds}
-}
-
-func (s *execWs) Connect(secret string, r *http.Request, w http.ResponseWriter) error {
-	for fd, fdSecret := range s.fds {
-		if secret == fdSecret {
-			conn, err := shared.WebsocketUpgrader.Upgrade(w, r, nil)
-			if err != nil {
-				return err
-			}
-
-			s.conns[fd] = conn
-
-			if fd == -1 {
-				s.controlConnected <- true
-				return nil
-			}
-
-			for i, c := range s.conns {
-				if i != -1 && c == nil {
-					return nil
-				}
-			}
-			s.allConnected <- true
-			return nil
-		}
-	}
-
-	/* If we didn't find the right secret, the user provided a bad one,
-	 * which 403, not 404, since this operation actually exists */
-	return os.ErrPermission
-}
-
-func runCommand(container *lxc.Container, command []string, options lxc.AttachOptions) shared.OperationResult {
-	status, err := container.RunCommandStatus(command, options)
-	if err != nil {
-		shared.Debugf("Failed running command: %q", err.Error())
-		return shared.OperationError(err)
-	}
-
-	metadata, err := json.Marshal(shared.Jmap{"return": status})
-	if err != nil {
-		return shared.OperationError(err)
-	}
-
-	return shared.OperationResult{Metadata: metadata, Error: nil}
-}
-
-func (s *execWs) Do() shared.OperationResult {
-	<-s.allConnected
-
-	var err error
-	var ttys []*os.File
-	var ptys []*os.File
-
-	if s.interactive {
-		ttys = make([]*os.File, 1)
-		ptys = make([]*os.File, 1)
-		ptys[0], ttys[0], err = shared.OpenPty(s.rootUid, s.rootGid)
-		s.options.StdinFd = ttys[0].Fd()
-		s.options.StdoutFd = ttys[0].Fd()
-		s.options.StderrFd = ttys[0].Fd()
-	} else {
-		ttys = make([]*os.File, 3)
-		ptys = make([]*os.File, 3)
-		for i := 0; i < len(ttys); i++ {
-			ptys[i], ttys[i], err = shared.Pipe()
-			if err != nil {
-				return shared.OperationError(err)
-			}
-		}
-		s.options.StdinFd = ptys[0].Fd()
-		s.options.StdoutFd = ttys[1].Fd()
-		s.options.StderrFd = ttys[2].Fd()
-	}
-
-	controlExit := make(chan bool)
-	stdEof := make(chan bool)
-
-	if s.interactive {
-		go func() {
-			select {
-			case <-s.controlConnected:
-				break
-
-			case <-controlExit:
-				return
-			}
-
-			for {
-				mt, r, err := s.conns[-1].NextReader()
-				if mt == websocket.CloseMessage {
-					break
-				}
-
-				if err != nil {
-					shared.Debugf("got error getting next reader %s", err)
-					break
-				}
-
-				buf, err := ioutil.ReadAll(r)
-				if err != nil {
-					shared.Debugf("failed to read message %s", err)
-					break
-				}
-
-				command := shared.ContainerExecControl{}
-
-				if err := json.Unmarshal(buf, &command); err != nil {
-					shared.Debugf("failed to unmarshal control socket command: %s", err)
-					continue
-				}
-
-				if command.Command == "window-resize" {
-					winch_width, err := strconv.Atoi(command.Args["width"])
-					if err != nil {
-						shared.Debugf("Unable to extract window width: %s", err)
-						continue
-					}
-
-					winch_height, err := strconv.Atoi(command.Args["height"])
-					if err != nil {
-						shared.Debugf("Unable to extract window height: %s", err)
-						continue
-					}
-
-					err = shared.SetSize(int(ptys[0].Fd()), winch_width, winch_height)
-					if err != nil {
-						shared.Debugf("Failed to set window size to: %dx%d", winch_width, winch_height)
-					}
-				}
-
-				if err != nil {
-					shared.Debugf("got error writing to writer %s", err)
-					break
-				}
-			}
-		}()
-
-		shared.WebsocketMirror(s.conns[0], ptys[0], ptys[0])
-	} else {
-		for i := 0; i < len(ttys); i++ {
-			go func(i int) {
-				if i == 0 {
-					<-shared.WebsocketRecvStream(ttys[i], s.conns[i])
-					ttys[i].Close()
-				} else {
-					<-shared.WebsocketSendStream(s.conns[i], ptys[i])
-					ptys[i].Close()
-					stdEof <- true
-				}
-			}(i)
-		}
-	}
-
-	result := runCommand(
-		s.container,
-		s.command,
-		s.options,
-	)
-
-	if !s.interactive {
-		ttys[0].Close()
-		ttys[1].Close()
-		ttys[2].Close()
-		<-stdEof
-	}
-
-	for _, tty := range ttys {
-		tty.Close()
-	}
-
-	for _, pty := range ptys {
-		pty.Close()
-	}
-
-	if s.interactive && s.conns[-1] == nil {
-		controlExit <- true
-	}
-
-	return result
-}
-
-type commandPostContent struct {
-	Command     []string          `json:"command"`
-	WaitForWS   bool              `json:"wait-for-websocket"`
-	Interactive bool              `json:"interactive"`
-	Environment map[string]string `json:"environment"`
-}
-
-func containerExecPost(d *Daemon, r *http.Request) Response {
-	name := mux.Vars(r)["name"]
-	c, err := newLxdContainer(name, d)
-	if err != nil {
-		return SmartError(err)
-	}
-
-	if !c.c.Running() {
-		return BadRequest(fmt.Errorf("Container is not running."))
-	}
-
-	post := commandPostContent{}
-	buf, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return BadRequest(err)
-	}
-
-	if err := json.Unmarshal(buf, &post); err != nil {
-		return BadRequest(err)
-	}
-
-	opts := lxc.DefaultAttachOptions
-	opts.ClearEnv = true
-	opts.Env = []string{}
-
-	if post.Environment != nil {
-		for k, v := range post.Environment {
-			if k == "HOME" {
-				opts.Cwd = v
-			}
-			opts.Env = append(opts.Env, fmt.Sprintf("%s=%s", k, v))
-		}
-	}
-
-	if post.WaitForWS {
-		ws := &execWs{}
-		ws.fds = map[int]string{}
-		if c.idmapset != nil {
-			ws.rootUid, ws.rootGid = c.idmapset.ShiftIntoNs(0, 0)
-		}
-		ws.conns = map[int]*websocket.Conn{}
-		ws.conns[-1] = nil
-		ws.conns[0] = nil
-		if !post.Interactive {
-			ws.conns[1] = nil
-			ws.conns[2] = nil
-		}
-		ws.allConnected = make(chan bool, 1)
-		ws.controlConnected = make(chan bool, 1)
-		ws.interactive = post.Interactive
-		ws.done = make(chan shared.OperationResult, 1)
-		ws.options = opts
-		for i := -1; i < len(ws.conns)-1; i++ {
-			ws.fds[i], err = shared.RandomCryptoString()
-			if err != nil {
-				return InternalError(err)
-			}
-		}
-
-		ws.command = post.Command
-		ws.container = c.c
-
-		return AsyncResponseWithWs(ws, nil)
-	}
-
-	run := func() shared.OperationResult {
-
-		nullDev, err := os.OpenFile(os.DevNull, os.O_RDWR, 0666)
-		if err != nil {
-			return shared.OperationError(err)
-		}
-		defer nullDev.Close()
-		nullfd := nullDev.Fd()
-
-		opts.StdinFd = nullfd
-		opts.StdoutFd = nullfd
-		opts.StderrFd = nullfd
-
-		return runCommand(c.c, post.Command, opts)
-	}
-
-	return AsyncResponse(run, nil)
-}
-
-var containerExecCmd = Command{name: "containers/{name}/exec", post: containerExecPost}
 
 /*
  * This is called by lxd when called as "lxd forkstart <container>"
